Server/requirements.txt
Flask==3.0.0
prometheus_client
python-keycloak
prometheus-flask-exporter==0.23.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
opentracing==2.4.0
requests==2.26.0
six==1.16.0
threadloop==1.0.2
thrift==0.13.0
tornado==6.1
urllib3==1.26.6